requirements.txt
tensorflow
scikit-learn
matplotlib
seaborn
numpy
pandas
kagglehub
pyyaml
